Refactor interpreter using modules
